% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_decline.R
\name{an_decline_server}
\alias{an_decline_server}
\title{Decline analysis module - server}
\usage{
an_decline_server(id, ctx)
}
\arguments{
\item{id}{Module id.}

\item{ctx}{Reactive case context.}
}
\value{
Reactive analysis result (or a list with \code{error}).
}
\description{
Decline analysis module - server
}
\keyword{internal}
