% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_decline.R
\name{an_decline_ui}
\alias{an_decline_ui}
\title{Decline analysis module - UI}
\usage{
an_decline_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Decline analysis module - UI
}
\keyword{internal}
