% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_admin_server.R
\name{tab_admin_server}
\alias{tab_admin_server}
\title{Admin tab - server}
\usage{
tab_admin_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context.}
}
\description{
Admin tab - server
}
\keyword{internal}
