% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_decisions_server.R
\name{tab_decisions_server}
\alias{tab_decisions_server}
\title{Decisions tab - server}
\usage{
tab_decisions_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context.}
}
\description{
Decisions tab - server
}
\keyword{internal}
