% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_mywork_server.R
\name{tab_mywork_server}
\alias{tab_mywork_server}
\title{My Work tab - server}
\usage{
tab_mywork_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context (see \code{\link[=app_server]{app_server()}}).}
}
\description{
My Work tab - server
}
\keyword{internal}
