% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_opportunity_server.R
\name{tab_opportunity_server}
\alias{tab_opportunity_server}
\title{Opportunity workspace - server}
\usage{
tab_opportunity_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context.}
}
\description{
Opportunity workspace - server
}
\keyword{internal}
