% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_pipeline_server.R
\name{tab_pipeline_server}
\alias{tab_pipeline_server}
\title{Pipeline tab - server}
\usage{
tab_pipeline_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context.}
}
\description{
Pipeline tab - server
}
\keyword{internal}
