% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stats_server.R
\name{tab_stats_server}
\alias{tab_stats_server}
\title{Process Stats tab - server}
\usage{
tab_stats_server(id, app)
}
\arguments{
\item{id}{Module id.}

\item{app}{Shared app context.}
}
\description{
Process Stats tab - server
}
\keyword{internal}
