% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_callout}
\alias{wt_callout}
\title{Callout box}
\usage{
wt_callout(title, ..., type = "info", icon = NULL)
}
\arguments{
\item{title}{Title text.}

\item{...}{Body content.}

\item{type}{\code{info}, \code{ok}, \code{warn}, \code{bad}, \code{neutral}.}

\item{icon}{Font Awesome icon name.}
}
\description{
Callout box
}
