% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{wt_can}
\alias{wt_can}
\title{Does a user hold a permission?}
\usage{
wt_can(user, permission, cfg)
}
\arguments{
\item{user}{Output of \code{\link[=wt_resolve_user]{wt_resolve_user()}}.}

\item{permission}{Permission id.}

\item{cfg}{Config list.}
}
\description{
Does a user hold a permission?
}
