% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_class_badge}
\alias{wt_class_badge}
\title{Badge for a complexity class}
\usage{
wt_class_badge(class, cfg)
}
\arguments{
\item{class}{Class id.}

\item{cfg}{Config list.}
}
\description{
Badge for a complexity class
}
