% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{wt_coerce_like}
\alias{wt_coerce_like}
\title{Coerce a value to the type of a template column}
\usage{
wt_coerce_like(x, template)
}
\description{
Coerce a value to the type of a template column
}
\keyword{internal}
