% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{wt_db_connect}
\alias{wt_db_connect}
\title{Connect to the WellTrack database}
\usage{
wt_db_connect(path = Sys.getenv("WT_DB_PATH", ":memory:"))
}
\arguments{
\item{path}{DuckDB file; \code{":memory:"} for the mockup. Defaults to env var \code{WT_DB_PATH}.}
}
\description{
Connect to the WellTrack database
}
