% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{wt_db_read_all}
\alias{wt_db_read_all}
\title{Read all tables into a named list}
\usage{
wt_db_read_all(con)
}
\arguments{
\item{con}{DBI connection.}
}
\description{
Read all tables into a named list
}
