% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{wt_db_seed}
\alias{wt_db_seed}
\title{Seed the database from a list of data frames (mockup / tests)}
\usage{
wt_db_seed(con, data)
}
\arguments{
\item{con}{DBI connection.}

\item{data}{Named list of data frames.}
}
\description{
Seed the database from a list of data frames (mockup / tests)
}
