% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_fmt}
\alias{wt_fmt}
\title{Format a number compactly}
\usage{
wt_fmt(x, suffix = "", digits = 0)
}
\arguments{
\item{x}{Number.}

\item{suffix}{Unit suffix.}

\item{digits}{Rounding.}
}
\description{
Format a number compactly
}
