% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_newopp.R
\name{wt_framing_input}
\alias{wt_framing_input}
\title{Build one framing input from its config definition}
\usage{
wt_framing_input(f, ns, cfg, wells)
}
\arguments{
\item{f}{Field definition.}

\item{ns}{Namespace function.}

\item{cfg}{Config list.}

\item{wells}{Known wells.}
}
\description{
Build one framing input from its config definition
}
\keyword{internal}
