% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_logic.R
\name{wt_gate_readiness}
\alias{wt_gate_readiness}
\title{Readiness percentage of a gate evaluation}
\usage{
wt_gate_readiness(eval)
}
\arguments{
\item{eval}{Output of \code{\link[=wt_evaluate_gate]{wt_evaluate_gate()}}.}
}
\description{
Readiness percentage of a gate evaluation
}
