% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wt_kpis}
\alias{wt_kpis}
\title{Value KPIs of the process}
\usage{
wt_kpis(data, dur, cfg)
}
\arguments{
\item{data}{Named list of tables.}

\item{dur}{Output of \code{\link[=wt_step_durations]{wt_step_durations()}}.}

\item{cfg}{Config list.}
}
\value{
Named list of KPI values.
}
\description{
Value KPIs of the process
}
