% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wt_load_config}
\alias{wt_load_config}
\title{Load the process configuration}
\usage{
wt_load_config(path = Sys.getenv("WT_CONFIG_PATH", ""))
}
\arguments{
\item{path}{YAML file. Defaults to env var \code{WT_CONFIG_PATH}, else the
packaged \code{inst/config/process.yml}.}
}
\value{
A validated list.
}
\description{
Load the process configuration
}
