% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_lookup}
\alias{wt_lookup}
\title{Safe lookup in a named vector}
\usage{
wt_lookup(x, key)
}
\arguments{
\item{x}{Named vector (or NULL).}

\item{key}{Name.}
}
\description{
Safe lookup in a named vector
}
\keyword{internal}
