% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R
\name{wt_mock_activity}
\alias{wt_mock_activity}
\title{Mock activity log (raw inputs already mapped to phases)}
\usage{
wt_mock_activity(cfg, opps, users, now)
}
\description{
Mock activity log (raw inputs already mapped to phases)
}
\keyword{internal}
