% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_items.R
\name{wt_my_work}
\alias{wt_my_work}
\title{Work items for a user (the "My Work" inbox)}
\usage{
wt_my_work(data, summary, user, cfg, now = Sys.time())
}
\arguments{
\item{data}{Named list of tables.}

\item{summary}{Output of \code{\link[=wt_portfolio_summary]{wt_portfolio_summary()}}.}

\item{user}{Output of \code{\link[=wt_resolve_user]{wt_resolve_user()}}.}

\item{cfg}{Config list.}

\item{now}{Reference time.}
}
\value{
data.frame(priority, kind, opp_id, well, item, due, age_days, action)
}
\description{
Exception-oriented: only items that need this person's action.
}
