% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{wt_permissions}
\alias{wt_permissions}
\title{Permissions granted by a set of roles}
\usage{
wt_permissions(roles, cfg)
}
\arguments{
\item{roles}{Character vector of role ids.}

\item{cfg}{Config list.}
}
\description{
Permissions granted by a set of roles
}
