% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_progress}
\alias{wt_progress}
\title{Thin progress bar}
\usage{
wt_progress(pct, label = NULL)
}
\arguments{
\item{pct}{Percentage 0-100.}

\item{label}{Optional label.}
}
\description{
Thin progress bar
}
