% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_logic.R
\name{wt_risk_score}
\alias{wt_risk_score}
\title{Risk score (probability x consequence, 1-5 scales)}
\usage{
wt_risk_score(probability, consequence)
}
\arguments{
\item{probability, consequence}{Numeric vectors.}
}
\description{
Risk score (probability x consequence, 1-5 scales)
}
