% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_logic.R
\name{wt_state_step}
\alias{wt_state_step}
\title{Step that a state belongs to}
\usage{
wt_state_step(state, cfg)
}
\arguments{
\item{state}{State id.}

\item{cfg}{Config list.}
}
\description{
Step that a state belongs to
}
