% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_status_chip}
\alias{wt_status_chip}
\title{Status chip with automatic colour}
\usage{
wt_status_chip(status)
}
\arguments{
\item{status}{Status keyword.}
}
\description{
Status chip with automatic colour
}
