% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{wt_status_type}
\alias{wt_status_type}
\title{Map a status keyword to a badge type}
\usage{
wt_status_type(status)
}
\arguments{
\item{status}{Status keyword.}
}
\description{
Map a status keyword to a badge type
}
