% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wt_step_stats}
\alias{wt_step_stats}
\title{Summary statistics of step durations}
\usage{
wt_step_stats(dur, by = "step_id", completed_only = TRUE)
}
\arguments{
\item{dur}{Output of \code{\link[=wt_step_durations]{wt_step_durations()}}.}

\item{by}{Grouping columns.}

\item{completed_only}{Use only finished step occurrences.}
}
\description{
Summary statistics of step durations
}
