% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wt_type_labels}
\alias{wt_type_labels}
\title{Named vector id -> label for intervention types}
\usage{
wt_type_labels(cfg)
}
\arguments{
\item{cfg}{Config list.}
}
\description{
Named vector id -> label for intervention types
}
