% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wt_workstreams}
\alias{wt_workstreams}
\title{Readiness workstreams as a data frame}
\usage{
wt_workstreams(cfg)
}
\arguments{
\item{cfg}{Config list.}
}
\description{
Readiness workstreams as a data frame
}
